Makefile: Adding CEPH_POINT_RELEASE support
This patch is about adding an option to enforce a particular point release of ceph like "12.2.0" instead of the latest.

This requires adding a field (named CEPH_POINT_RELEASE) on the flavor name, just after the CEPH_VERSION field so we have something like :
    luminous,-12.2.2,[....]

This change implies this CEPH_POINT_RELEASE is encoding the release but also the syntax for a specific package manager.
For centos, yum expect the package version to be prefixed by a dash, so CEPH_POINT_RELEASE should have a dash for centos.
This make possible having a generic patch for all distro.

Note that if the field is let empty, like in "luminous,,x86_64,[...]", the field will be automatically replaced by an empty string.
This is very useful for users not to have the need of writing "luminous,'',x86_64,[...]"
An empty CEPH_POINT_RELEASE, which is the current default, leads to installing the lastest available version of the ceph packages.

This patch only implements it for centos as it will depends on distributions & repositories to offer/enable this feature.
Signed-off-by: Erwan Velu <evelu@redhat.com>